Move side filters outside dashboard body
